Add support for Gloebits
- Adds support for Gloebits as an addon-module

NOTE: The Second Galaxy Development Team does not endorse nor can we provide support for Gloebits.  We only provide the minimal support to get it building in Virtual Universe.  It is the responsibility of the end user to use the gloebits virtual world currency service in a responsible manner and to adhere to all applicable laws of the jurisdiction in which you live.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>