Adjust the timeout for asset uploads
- Correct licensing text to pass our security scan tools
- Organize using references
- Increase the timeout for asset uploads

NOTE: This is a temporary fix until we can come up with a better fix.  It is intended to ensure that uploads of assets that are larger in total size will properly upload and store.  It should be note that this matches the timeout for asset uploads on Second Galaxy via our proprietary grid architecture.

NOTE 2: Just because depreciating a cap seems like a good idea does not mean that it is.  OpenSim devs have a habit of making changes without fully testing or thinking these things through very carefully.  This has been evident for quite a while.

Signed-Off-By: NoahStarfinder<noah@secondgalaxy.com>