Disable MeshModelAllowTextureToInventory by default
- Disable MeshModelAllowTextureToInventory by default

NOTE: By enabling MeshModelAllowTextureToInventory you run the risk of flooding your inventory and asset services with unnecessary assets and inventory items given how many times you might upload a mesh model before your satisfied with it.  This would cause clutter and make everyone's inventory much more of a mess then really is necessary.  Textures should already be on the model before uploading the model.

This will be a matter for each individual grid owner to decide for themselves whether to enable it or not.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>