[Experimental]: Add a BaseWebURL constants set to the configuration files
- Adds a constants set to the [Const] section of the Robust and OpenSim configuration files
- Updates the urls for various services that run on your grid's main website server

NOTE: This is an experiment to see if this in fact will work.  You do not need to update your version of Virtual Universe if you choose not to.  This test is being done prior to an upcoming major commit to see if it will work and if it will help make things a bit easier on people.  It also cleans up the explanation a bit of what the urls in the const section actually do.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>