Commit fresh copy of codebase that will be in the base of Virtual-Universe-NXG from now on.
NOTE 1:  For those who are wanting to use Virtual Universe to connect to HyperGrid, At the moment this is not recommended as it is not compatible.

NOTE 2: Do not switch back and forth between OpenSimulator Version 0.9.x and Virtual Universe.  Doing so will corrupt your data and break your database tables.

NOTE 3: There will be many changes over the next weeks and months.  So do not expect us to provide complete compatibility with OpenSimulator as that is not our intent.  We will provide as much compatibility as we deem appropriate.

Signed-Off-By: EmperorStarfinder<emperor@secondgalaxy.com>