Clean code
- Many thanks to the wonderful folks from the MOSES team for their hard work.  Your work will live on in Virtual Universe

NOTE: This is Virtual Universe 2.001 - Dev.

You should be aware that this release is not necessarily fully compatible with OpenSimulator or Hypergrid.  It is not intended to be compatible fully with either OpenSimulator or it's version of Hypergrid.  It may however still work and we will do our best to maintain compability where time is available to us.  By using this version of our architecture you understand this real hard truth and fact.

This architecture is the open source sister architecture of our proprietary architecture that runs Second Galaxy.  As we progress you will see us move this version toward a better architecture for all virtual worlds, whether it be a full commercial level grid, or just a sandbox to play around in.

Note 2: Until our physics engine developer commits the built binary files for PhysX please use BulletSim.

Signed-Off-By: EmperorStarfiner<emperor@secondgalaxy.com>