Refactor VersionInfo
- Refactor VersionInfo to work more like a commercial grid similar to Secondlife

NOTE 1: This change may affect users ability to use HyperGrid when attempting to connect from a region or grid on Virtual Universe to a grid or region running on OpenSimulator.

NOTE 2: With this commit the version number of all releases will show the number of days since 2000 as the revision number based upon the date of the build.  This is similar to how it is done in Secondlife.

Signed-Off-By: BruceStarfinder<bruce@secondgalaxy.com>