Begin reworking of Hyper Grid SuitCase Inventory Service API
- Remove a bug known to OpenSimulator developers in 2016 but neglected to fix
- Fix licensing text
- Begin reworking to meet our new API requirements
- Add curly brackets to necessary statements to ensure compiler tools exit statements in a clean fashion
- Remove commented out code to help prevent bloating in built binary files
- Clenup comments
- Cleanup console messages
- Cleanup whitespaces
- Cleanup Trailing line endings

NOTE: Virtual Universe may or may not work with Hypergrid at the moment while we work to replace the current API with our API for Spacegate.  Once this piece has been completed we will hook up the necessary compatibility modules to handle users coming to grids on our architecture from Hypergrid and Spacegate.
Signed-off-by: CaseyraeStarfinder <caseyrae@secondgalaxy.com>