Begin putting the brand new file diretory structure in place
- Add the file directory structure so it looks neat and professionally clean
- Add the AGPL license.  We now use the AGPL license for the core architecture
- Add placeholders in each folder to maintain the folder in the repository

NOTE: We have chosen to write the architecture from the ground up for many reasons.  A big reason is that there are far to many bugs, security issues, and unprofessional and nonsensical code written in OpenSimulator that causes their architecture to be far to clumbsy when running a commercial grade virtual world.  Unforunately most of the forks of OpenSimulator while some have made a valiant attempt to correct the problems did not go deep or far enough to fully fix the numerous issues under the hood.  This includes all forks from early OpenSimulator days dating back at the very minimum to 2008.  Grid owners deserve better and our goal is to provide them with the best quality possible while empowering the future of virtual worlds.
Signed-off-by: NoahStarfinder <noah@secondgalaxy.com>